dynamic 2d window zoom
Complemented the uniform 2x zoom by adding a drag+drop
to zoom/cancel zoom on a specific region.
This triggered a possible engine issue where rendertargets
are not being correctly flagged dirty when we change texture
coordinates on something that is being composed (easiest
to see with histogram on zoom region 'lagging').