enabled per/block transfer patterns in alpha
With ARCAN_ARG=val:0x11,0x22:opt= style env. argument
passing, a number of patterns can be added. When the
output is packed (and alpha channel mode is set to PTN),
the buffer bytes that match this pattern will have the
pattern # encoded into the alpha channel of the output.