merged translator feature
This is the major change for the next (0.2) version
of senseye, 'translators'. A translator is a domain-
specific decoder that generates a higher, more abstract
view of a slice of data. While sensors provide raw
data streams and senseye UI then provides visual
representations and statistical tools for finding
interesting data blocks, translators then close the
loop by providing more specific analysis on a target by
target basis.

Another side-case is that the translator feature can be
used as a scratch-pad for developing a parser while
reversing and get a very tight feedback loop.

The primary included translators are:
 xlt_ascii - text/7-bit ascii rendering
 xlt_capstone - disassembly view

and upcoming ones should address hex-view, audio and
image formats.

There are still some minor fixes and tuning to fix in
the individual translators, but the branch is mature enough
to be merged into master.