datawnd popup translator fix
previously, the translator feature was only available through
rather obscure keycommands, (meta+rclick or meta+shift+T). This
commit moves that feature to the more suitable Translation ..
entry in the regular popup. This hides the offset on mouse motion
feature, but there is a more useful highlight for that coming.