added non-uniform scale, some issues though
When meta- moving over a supported surface, the lower 1/3 corner
turns into a resize icon. This works correctly for the data window
(as that is just a scaling operation) but dragging a translator
window requires that shared memory buffers etc. are recalculated
which is a highly asynchronous operation. This highlighted a bug
in arcan that does not calculate scaling factors for this correctly.
Until that has been fixed, the resize on translators is a bit hacky;
translator controls output dimensions, hint is sent on explicit
uniform changes (F2/META+F2) and on mouse drop and otherwise just
stretched when dragging.