first sketch of pictune added
Pictune is to be used for finding raw image buffers in sensor- provided
data. It provides a stage for color space unpacking, one for stride
tuning and one for offset alignment and begin / end detection.

This commit is rather crude with most of those features incomplete,
although they will be fleshed out in the coming few pushes.