first push towards overlay support
Overlays are intended to be a feature for more advanced translators and
for the next 'integrator' feature (sensor and translator in one, for
integration with other complex tools).

The idea is to have short-lived translator subsegments that can be
toggled by the user to be drawn at some level of transparency above the
regular data window. Input will behave like it normally would in the
data window, but the overlay would also receive it.

This is a stepping stone towards overlaying color and text data on top
of the data window for things such as allocated blocks or symbol
references.