simplified xlt_img somewhat
most desired features for this release of xlt_img are done, the one
thing missing still is the current overlay bug where, if you
"seek-click", the offset and region that the overlay colors is broken.
In addition, the overlay does not handle list mode at all (should be
trivial to fix).